Consolidate worker to monolithic version
